Add a ttl option for s3 objects
This commit allows the user to optionally specify a ttl (time to live)
flag in seconds for s3 objects, optionally on a specific bucket.
Objects older than the ttl are reaped.

This is necessary since should many gigabytes be written to s3, a
built-in way to clear up this data is much more efficient than making
many IO calls.